+++ libraries/MSP300/MSP300.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: MSP300.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.3.0
 // PURPOSE: Arduino library for I2C MSP300 pressure transducer.
 //     URL: https://github.com/RobTillaart/MSP300
 
@@ -11,11 +11,14 @@
 
 MSP300::MSP300(const uint8_t deviceAddress, TwoWire *wire)
 {
-  _address    = deviceAddress;
-  _wire       = wire;
-  _error      = MSP300_OK;
-  _maxValue   = 100;
-  _raw        = 0;
+  _address     = deviceAddress;
+  _wire        = wire;
+  _error       = MSP300_OK;
+  _maxValue    = 100;
+  _pressure    = 0;
+  _temperature = 0;
+  _status      = MSP300_OK;
+  setPressureCounts(1000, 15000);
 }
 
 
@@ -52,9 +55,21 @@ uint8_t MSP300::getAddress()
 //
 //  CALIBRATION
 //
-void MSP300::setPressureCounts(uint16_t Pmin, uint16_t Pmax)
+void MSP300::setPressureCounts(int Pmin, int Pmax)
 {
-  _pressureFactor = 1.0 / (Pmax - Pmin);
+  _Pmin = Pmin;
+  _Pmax = Pmax;
+  _pressureFactor = _maxValue / (_Pmax - _Pmin);
+}
+
+int MSP300::getPmin()
+{
+  return _Pmin;
+}
+
+int MSP300::getPmax()
+{
+  return _Pmax;
 }
 
 
@@ -62,35 +77,53 @@ void MSP300::setPressureCounts(uint16_t Pmin, uint16_t Pmax)
 //
 //  READ
 //
+uint32_t MSP300::readP()
+{
+  _request();
+  //  read status + pressure
+  uint32_t raw = _read(2);
+
+  int P = raw && 0x3FFF;   //  14 bit
+  _pressure = (P - _Pmin) * _pressureFactor;
+
+  _status = (raw >> 14);  //  2 bit
+
+  return raw;
+}
+
+
 uint32_t MSP300::readPT()
 {
   _request();
-  _read(4);    //  read all data.
-  //  TODO check status
-  //  uint8_t status = (_raw >> 30);
-  //  00 = OK
-  //  01 = reserved
-  //  10 = stale data
-  //  11 = error
-  return _raw;
+  //  read status, pressure and temperature
+  uint32_t raw = _read(4);
+
+  int T = (raw >> 5) && 0x07FF;   //  11 bit
+  _temperature = T * (200.0 / 2048) - 50.0;
+
+  int P = (raw >> 16) && 0x3FFF;  //  14 bit
+  _pressure = (P - _Pmin) * _pressureFactor;
+
+  _status = (raw >> 30);         //  2 bit
+
+  return raw;
 }
 
+
+uint8_t MSP300::getStatus()
+{
+  return _status;
+}
 
 float MSP300::getPressure()
 {
-  //  formula page 5 datasheet
-  int pres = (_raw >> 16) && 0x3FFF;
-
-  return pres * (_maxValue * _pressureFactor);
+  return _pressure;
 }
-
 
 float MSP300::getTemperature()
 {
-  int temp = (_raw >> 5) && 0x07FF;
-  return temp * (200.0 / 2048) - 50.0;
+  return _temperature;
 }
-
 
 int MSP300::lastError()
 {
@@ -111,19 +144,21 @@ void MSP300::_request()
 }
 
 
-void MSP300::_read(uint8_t bytes)
+uint32_t MSP300::_read(uint8_t bytes)
 {
   if (_wire->requestFrom(_address, bytes) != bytes)
   {
-    _error = MSP300_ERROR;
-    return;  //  keep last value
+    _error = MSP300_REQUEST_ERROR;
+    //  report a read error.
+    return (uint32_t(MSP300_READ_ERROR) << 30);
   }
-  _raw = 0;
+  uint32_t raw = 0;
   for (int i= 0; i < bytes; i++)
   {
-    _raw <<= 8;
-    _raw = _wire->read();
+    raw <<= 8;
+    raw |= _wire->read();
   }
+  return raw;
 }