+++ libraries/MSP300/CHANGELOG.md
@@ -6,6 +6,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - 2024-05-09
+- Fix read bug, thanks to Gabo-Bravo
+- updated readme.md
+- added **uint8_t getStatus()** + defines for error codes.
+- update the default I2C address to 0x28 (within range).
+- rewrite **readPT()** and cache pressure, temperature and status.
+- add **readP()** to read pressure and status only. (faster).
+- add **getPmin()** and **getPmax()** for calibration.
+- add **MSP300_performance.ino** to measure I2C read performance.
+- add **MSP300_demo_pressure_only.ino** 
+- add **MSP300_fast_pressure.ino** 
+- update keywords.txt
+
+----
+
 ## [0.2.0] - 2023-12-05
 - refactor begin()
 - update readme.md